homomorphic/bin/requirements.txt
pyfhel==3.1.4
numpy==1.23.0
splunklib
splunk-sdk
python-dotenv
